Tweak logic for parsing and setting IDs for `ContentFeed` and `ContentFeedItem`.
For the feed items, the items will still have a INT value as the ID, but it will be returned as a string from server and we should store it as a string (since the new FeedItem model will have a string id).

However, hypothetically, we should be able to convert the string to an int as needed.